Création du formlaire fonctionnel pour la modification de demande spéciale
